Fixed FactoryGirl issuea
 - non-string value is not allowed
 - exception should be thrown when variable is unset

+++ factory-girl/src/test/scala/blog/Post.scala
@@ -8,7 +8,7 @@ case class Post(
   id: Long,
   title: String,
   body: String,
-  viewCount: BigDecimal,
+  viewCount: Long,
   tags: Seq[Tag] = Nil,
   createdAt: DateTime,
   updatedAt: Option[DateTime] = None)
